MongoosePush v2.0.1 MongoosePush.Config.ConfexProvider

Functions

init(opts)

init([any()]) :: :ok

Called when the provider is initialized.

Providers are invoked pre-boot, in a dedicated VM, with all application code loaded, and kernel, stdlib, compiler, and elixir applications started. Providers must use this callback to push configuration into the application environment, which will be persisted to a final sys.config for the release itself.

The arguments given to init/1 are the same as given in the config_providers setting in your release configuration file.

Callback implementation for Distillery.Releases.Config.Provider.init/1.