MongoosePush v2.0.1 MongoosePush.Metrics

This module provides some utility macros that simplify use of Elixometer.

Functions

name(type, prefix, suffix)
(macro)

update(return_value, mtype, metric, value \\ 1)
(macro)

Updates metric (spiral) by given value. The metrics name is generated based on given prefix and the return value of the tested function. Provided return value of :ok is counted as succeses, while {:error, reason :: term} as error reason.

update_metric(atom, metric, value)
(macro)