MongoosePush v2.1.0-alpha.1 MongoosePush.Service.FCM.Pools

This module is responsible for worker pools management. It provides several utility functions that help with e.g. selecting workers for given pool of the service.

Functions

pool_size(service, name)

Specs

pool_size(MongoosePush.service(), atom()) :: integer()

Returns size of the pool

pools_by_mode()

Specs

pools_by_mode() :: [atom()]

Returns lists of pool names that have selected :mode set

select_worker()

Specs

select_worker() :: atom()

Return random worker name for given service and with given :mode set

worker_name(type, name, num)

Specs

worker_name(atom(), atom(), integer()) :: atom()

Returns worker name based of the service type, worker name and its id